{{cookiecutter.project_name}}/requirements.txt
flask
flask-script
flask-migrate
flask-sqlalchemy
graphene
